+++ packages/anon-aadhaar-pcd/src/pcd.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/ban-ts-comment */
 import { DisplayOptions, PCD, PCDPackage, SerializedPCD } from '@pcd/pcd-types'
 import {
   PCDInitArgs,
@@ -9,10 +8,7 @@ import {
 } from './types'
 
 import { v4 as uuidv4 } from 'uuid'
-
-// @ts-ignore
 import { groth16 } from 'snarkjs'
-
 import { splitToWords } from './utils'
 import JSONBig from 'json-bigint'
 import { BackendProver, ProverInferace, WebProver } from './prover'